break: rework session handling (#7)
THIS IS A BREAKING CHANGE!!!

Sessions are now stored in the cache dir of your OS.

The sessions are stored in their own directory that is the
SHA256 checksum of the absolute project_root dir.

Additonally metadata.json is also stored in the session dir.

The metadata.json contains only the project_dir for now.

This makes it easier to get the plugin up and running without
messing around with repository directory structures.

Also you don't need to add the .editor/ directory to the vsc ignore
anymore.