fix(session): load conflicting with nvim start cwd
Let's say I have a git project in ~/projects/personal/foo.

This project contains a direct child-folder bar.

If I change to ~/projects/personal/foo/bar and open neovim the root cwd for neovim is the folder foo,
which is an expected behaviour.

Now I open some files and quit neovim. Kikao persists them via session.nvim in the kikao.nvim project cache folder.

If I know enter ~/projects/personal/foo and open neovim,
the cwd changes due to sourcing the session, but this is wrong.
We want to keep the current cwd.

Before sourcing, we store the current cwd,
then try to restore the session and then change back the cwd.